refactor: 패키지 역할에 맞게 파일 위치 이동 & 하드 코딩 에러 메세지 클래스화
