+++ geulnamu-backend/src/main/java/com/geulnamu/controller/member/MemberController.java
@@ -4,10 +4,9 @@ import com.geulnamu.controller.member.dto.request.*;
 import com.geulnamu.controller.member.dto.response.MemberInfoResponseDTO;
 import com.geulnamu.controller.member.dto.response.MemberListResponseDTO;
 import com.geulnamu.domain.shared.enums.Level;
-import com.geulnamu.domain.shared.paging.PagingRequest;
-import com.geulnamu.global.response.BaseResponse;
+import com.geulnamu.infrastructure.response.paging.PagingRequest;
+import com.geulnamu.infrastructure.response.BaseResponse;
 import com.geulnamu.infrastructure.annotation.AuthMemberId;
-import com.geulnamu.infrastructure.annotation.AuthToken;
 import com.geulnamu.infrastructure.annotation.AccessLevel;
 import com.geulnamu.service.member.MemberService;
 import jakarta.validation.Valid;
