+++ geulnamu-backend/src/main/java/com/geulnamu/infrastructure/config/security/AuthTokenResolver.java
@@ -1,7 +1,7 @@
 package com.geulnamu.infrastructure.config.security;
 
-import com.geulnamu.domain.shared.TokenInfo;
-import com.geulnamu.global.response.ResponseMessage;
+import com.geulnamu.infrastructure.security.token.TokenInfo;
+import com.geulnamu.infrastructure.response.ResponseMessage;
 import com.geulnamu.infrastructure.annotation.AuthToken;
 import com.geulnamu.infrastructure.exception.TokenException;
 import io.micrometer.common.util.StringUtils;
