+++ geulnamu-backend/src/main/java/com/geulnamu/infrastructure/exception/NotFoundDataException.java
@@ -1,6 +1,6 @@
 package com.geulnamu.infrastructure.exception;
 
-import com.geulnamu.global.response.ResponseMessage;
+import com.geulnamu.infrastructure.response.ResponseMessage;
 import lombok.Getter;
 
 @Getter
