+++ geulnamu-backend/src/main/java/com/geulnamu/infrastructure/response/paging/PagingResponse.java
@@ -1,4 +1,4 @@
-package com.geulnamu.domain.shared.paging;
+package com.geulnamu.infrastructure.response.paging;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
