+++ geulnamu-backend/src/main/java/com/geulnamu/infrastructure/security/token/TokenPair.java
@@ -1,4 +1,4 @@
-package com.geulnamu.domain.shared;
+package com.geulnamu.infrastructure.security.token;
 
 public record TokenPair(String accessToken, String refreshToken) {
 }
