+++ geulnamu-backend/src/main/java/com/geulnamu/infrastructure/security/token/TokenType.java
@@ -1,4 +1,4 @@
-package com.geulnamu.domain.shared.enums;
+package com.geulnamu.infrastructure.security.token;
 
 public enum TokenType {
     AccessToken,
