+++ geulnamu-backend/src/main/java/com/geulnamu/infrastructure/util/JwtTokenUtil.java
@@ -1,9 +1,9 @@
 package com.geulnamu.infrastructure.util;
 
 import com.geulnamu.domain.shared.enums.Role;
-import com.geulnamu.domain.shared.TokenInfo;
-import com.geulnamu.domain.shared.enums.TokenType;
-import com.geulnamu.global.response.ResponseMessage;
+import com.geulnamu.infrastructure.security.token.TokenInfo;
+import com.geulnamu.infrastructure.security.token.TokenType;
+import com.geulnamu.infrastructure.response.ResponseMessage;
 import com.geulnamu.infrastructure.exception.TokenException;
 import io.jsonwebtoken.*;
 import lombok.extern.slf4j.Slf4j;
