+++ geulnamu-backend/src/main/resources/static/docs/login-api-guide.html
@@ -513,7 +513,7 @@ Content-Type: application/x-www-form-urlencoded</code></pre>
 <div class="content">
 <pre class="highlightjs highlight"><code class="language-http hljs" data-lang="http">HTTP/2.0 200 OK
 Content-Type: application/json;charset=UTF-8
-Set-Cookie: refreshToken=random_refreshToken_code_2; Path=/; Max-Age=15552000; Expires=Thu, 04 Dec 2025 18:08:00 GMT; Secure; HttpOnly
+Set-Cookie: refreshToken=random_refreshToken_code_2; Path=/; Max-Age=15552000; Expires=Thu, 04 Dec 2025 19:34:08 GMT; Secure; HttpOnly
 
 {
   "code" : 200,
@@ -745,7 +745,7 @@ Host: docs.api.com</code></pre>
 <div class="content">
 <pre class="highlightjs highlight"><code class="language-http hljs" data-lang="http">HTTP/2.0 200 OK
 Content-Type: application/json;charset=UTF-8
-Set-Cookie: refreshToken=random_refreshToken_code; Path=/; Max-Age=15552000; Expires=Thu, 04 Dec 2025 18:07:59 GMT; Secure; HttpOnly
+Set-Cookie: refreshToken=random_refreshToken_code; Path=/; Max-Age=15552000; Expires=Thu, 04 Dec 2025 19:34:08 GMT; Secure; HttpOnly
 
 {
   "code" : 200,
