+++ geulnamu-backend/src/test/java/com/geulnamu/controller/member/LoginControllerTest.java
@@ -3,8 +3,8 @@ package com.geulnamu.controller.member;
 import com.geulnamu.controller.auth.LoginController;
 import com.geulnamu.controller.auth.dto.response.LoginResponseDTO;
 import com.geulnamu.controller.shared.ControllerTest;
-import com.geulnamu.domain.shared.TokenInfo;
-import com.geulnamu.global.response.ResponseMessage;
+import com.geulnamu.infrastructure.security.token.TokenInfo;
+import com.geulnamu.infrastructure.response.ResponseMessage;
 import com.geulnamu.service.auth.LoginFacade;
 import jakarta.servlet.http.Cookie;
 import org.junit.jupiter.api.Test;
