+++ geulnamu-backend/src/test/java/com/geulnamu/controller/member/MemberControllerTest.java
@@ -8,9 +8,9 @@ import com.geulnamu.controller.member.dto.response.MemberListResponseDTO;
 import com.geulnamu.controller.shared.ControllerTest;
 import com.geulnamu.domain.shared.enums.Gender;
 import com.geulnamu.domain.shared.enums.Role;
-import com.geulnamu.domain.shared.paging.PagingRequest;
-import com.geulnamu.domain.shared.paging.PagingResponse;
-import com.geulnamu.global.response.ResponseMessage;
+import com.geulnamu.infrastructure.response.paging.PagingRequest;
+import com.geulnamu.infrastructure.response.paging.PagingResponse;
+import com.geulnamu.infrastructure.response.ResponseMessage;
 import com.geulnamu.service.member.MemberService;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
